Migrate to new .fsproj format
Update 5 for Visual Studio 2017 shipped a few days after I started this year's challenge, and it changed the F# project templates from the old-style to the new .NET Core-style templates.

This new style is much simpler, and does a much better job of managing transitive NuGet dependencies, so I decided it was preferable to just move the first few projects over to the newer style. It avoids the need to add the same FSharp.Core and Unquote refs to every project, because they now get picked up from Common.